% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pal.R
\name{create_pal}
\alias{create_pal}
\title{Create a palette using colors extracted from an image}
\usage{
create_pal(
  image,
  n,
  resize = 0.1,
  method = "kmeans",
  colorblind = T,
  sort = "none",
  show.pal = T,
  title,
  ...
)
}
\arguments{
\item{image}{a path to the raster image (JPG, JPEG, PNG, TIFF) from which the
colors are to be extracted.}

\item{n}{a positive integer. The number of colors in the palette.}

\item{resize}{a number between 0 and 1. This indicates the fraction by which
the width/height (pixels) of the original image is resized while keeping
the aspect ratio. Default to 0.1.}

\item{method}{the clustering method for classifying pixels into groups based
on the RGB values. Options are \code{"kmeans"} and \code{"Gaussian_mix"}.
Default to \code{"kmeans"}.}

\item{colorblind}{logical. Whether to render the palette colorblind-friendly
(also see the ‘Details’ section). Default to \code{TRUE}.}

\item{sort}{a character indicating how the colors should be sorted. Options
are \code{"none"}, \code{"hue"}, \code{"saturation"}, and \code{"value"}.
Default to \code{"none"}(unsorted).}

\item{show.pal}{logical. Whether to display the palette or not. Default to \code{TRUE}.}

\item{title}{a character string giving the title of the displayed palette.}

\item{...}{additional arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme}}.}
}
\value{
A vector of hexadecimal color codes.
}
\description{
\code{create_pal} creates a colorblind-friendly palette using
    colors extracted from an image.
}
\details{
Two clustering methods are available. For \code{method = "kmeans"},
    pixels are partitioned into clusters using \code{\link[stats::kmeans]{kmeans}},
    and the RGB values of the cluster centers are converted into the corresponding
    hexadecimal color codes. For \code{method = "Gaussian_mix"}, pixel components
    are identified via multivariate Gaussian mixture modeling using \code{\link[ClusterR:GMM]{ClusterR::GMM}},
    and the RGB values of the component centroids are converted into the
    corresponding hexadecimal color codes.

    If \code{"colorblind = T"}, the original colors are replaced with
    colorblind-friendly colors using \code{\link[colorBlindness:replacePlotColor]{colorBlindness::replacePlotColor}}.

    The colors in the palette can be sorted based on their HSV values.
    If \code{sort = "hue"}, the colors are sorted by hue in an ascending order.
    If \code{sort = "saturation"}, the colors are sorted by saturation in a descending order.
    If \code{sort = "value"}, the colors are sorted by value in a descending order.
}
\examples{
\dontrun{
library(PalCreatoR)
image_path <- system.file("Mountain.JPG", package = "PalCreatoR")

My_pal <- create_pal(image = image_path,
                     n = 5,
                     resize = 0.1,
                     method = "kmeans",
                     colorblind = T,
                     sort = "value",
                     show.pal = T,
                     title = "My Palette")

print(My_pal)}
}
